third_party/WebKit/LayoutTests/http/tests/serviceworker/cache-match-expected.txt

This is a testharness.js-based test.
PASS Service Worker: Cache.match and Cache.matchAll 
FAIL Cache.matchAll with URL Method is not implemented.
FAIL Cache.match with URL assert_equals: Cache.match should match by URL.: Response.url expected "" but got "http://example.com/a"
FAIL Cache.matchAll with Request Method is not implemented.
FAIL Cache.match with Request assert_equals: Cache.match should match by Request.: Response.url expected "" but got "http://example.com/a"
FAIL Cache.matchAll with new Request Method is not implemented.
FAIL Cache.match with new Request assert_equals: Cache.match should match by Request.: Response.url expected "" but got "http://example.com/a"
PASS Cache.match with Request containing non-empty body 
FAIL Cache.matchAll with ignoreSearch option (request with no search parameters) Method is not implemented.
FAIL Cache.matchAll with ignoreSearch option (request with search parameter) Method is not implemented.
FAIL Cache.matchAll with request containing hash Method is not implemented.
FAIL Cache.matchAll with string fragment "http" as query Method is not implemented.
FAIL Cache.matchAll with prefixMatch option Method is not implemented.
FAIL Cache.matchAll with prefixMatch option Method is not implemented.
FAIL Cache.matchAll with responses containing "Vary" header assert_unreached: Test setup failed: Entry already exists. Reached unreachable code
FAIL Cache.match with responses containing "Vary" header assert_unreached: Test setup failed: Entry already exists. Reached unreachable code
FAIL Cache.matchAll with "ignoreVary" parameter assert_unreached: Test setup failed: Entry already exists. Reached unreachable code
FAIL Cache.match with Request and Response objects with different URLs assert_equals: Cache.match should reutrn a Response object that has the same properties as the stored response.: Response.url expected "http://127.0.0.1:8000/serviceworker/resources/simple.txt" but got "http://example.com/"
PASS Cache.match invoked multiple times for the same Request/Response 
Harness: the test ran to completion.
